+++ game/tests/test_views_master.py
@@ -9,6 +9,7 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 
+import chat.models as cmodels
 import game.forms as gforms
 import game.models as gmodels
 import game.views.master as gvmaster
@@ -74,6 +75,7 @@ class InviteCharacterViewTest(TestCase):
         user.save()
 
         game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         number_of_characters_with_game = 5
         number_of_characters_without_game = 12
         for i in range(number_of_characters_with_game):
@@ -170,7 +172,8 @@ class InviteCharacterConfirmViewTest(TestCase):
         user.user_permissions.add(permission)
         user.save()
 
-        gmodels.Game.objects.create()
+        game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         gmodels.Character.objects.create(name=utils.generate_random_name(5))
 
     def setUp(self):
@@ -233,7 +236,8 @@ class StartGameViewTest(TestCase):
         user.user_permissions.add(permission)
         user.save()
 
-        gmodels.Game.objects.create()
+        game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
 
     def setUp(self):
         self.user = User.objects.last()
@@ -360,6 +364,7 @@ class CreateTaleViewTest(TestCase):
         user.save()
 
         game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         game.start()
@@ -442,6 +447,7 @@ class CreatePendingActionViewTest(TestCase):
         user.save()
 
         game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         game.start()
@@ -569,6 +575,7 @@ class IncreaseXpViewTest(TestCase):
         user.save()
 
         game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         game.start()
@@ -693,6 +700,7 @@ class DamageViewTest(TestCase):
         user.save()
 
         game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         gmodels.Character.objects.create(game=game, name=utils.generate_random_name(5))
         game.start()
@@ -842,6 +850,7 @@ class HealViewTest(TestCase):
         user.save()
 
         game = gmodels.Game.objects.create()
+        cmodels.Room.objects.create(game=game)
         gmodels.Character.objects.create(
             game=game, name=utils.generate_random_name(5), hp=1
         )