+++ role_play/prod_settings.py
@@ -42,13 +42,16 @@ SECURE_HSTS_PRELOAD = True
 # Application definition
 
 INSTALLED_APPS = [
+    "daphne",
     "game.apps.GameConfig",
+    "chat.apps.ChatConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    "channels",
 ]
 
 MIDDLEWARE = [
@@ -80,7 +83,19 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "role_play.wsgi.application"
+# Channels
+
+
+ASGI_APPLICATION = "role_play.asgi.application"
+
+CHANNEL_LAYERS = {
+    "default": {
+        "BACKEND": "channels_redis.core.RedisChannelLayer",
+        "CONFIG": {
+            "hosts": [(os.environ["REDISHOST"], os.environ["REDISPORT"])],
+        },
+    },
+}
 
 
 # Database