+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+* Use Redis as cache
 * Various UX/UI improvements
 * Add index on event model
 * Upgrade project's libraries to their latest versions